Até a b da 4

+++ TP6/src/main2.cpp
@@ -1,5 +1,5 @@
 /* Author: Lauren Rolan
- * File: main1.cpp
+ * File: main2.cpp
  */
 
 #include <iostream>
@@ -10,10 +10,10 @@ using namespace std;
 int main(int, char *[])
 {
     /* Exercice 2 -- taille et capacite d'un vecteur */
-    vector<int> v;
+    vector<double> v;
     for(int i = 0; i < 3000; i++)
     {
-        v.push_back(i);
+        v.push_back((double)i);
         cout << "It: " << i << "  Cap: " << v.capacity() << endl;
     }
 }
